move app bar out of product folder
