+++ features/app-bar/settings-section.tsx
@@ -2,8 +2,8 @@
 import { Eraser, Save, Undo2, X } from "lucide-react";
 import { motion } from "motion/react";
 import { ButtonHTMLAttributes, useState } from "react";
-import { DescriptionSystemPrompt } from "../recommendation/prompt";
-import { updateSystemPrompt, useImageDescription } from "../store";
+import { DescriptionSystemPrompt } from "../product-page/recommendation/prompt";
+import { updateSystemPrompt, useImageDescription } from "../product-page/store";
 import { closeAppBarSection } from "./app-bar-store";
 
 export const SettingsSection = () => {
